+++ uis/pauseui.js
@@ -3,21 +3,18 @@ import { controlsScreen } from './controlsScreen.js'
 import { optionsScreen } from './options.js'
 import { serverlist } from './serverlist.js'
 import texts from '../js/lang.js'
-import { LocalSocket } from '../js/iframe.js'
 import { worldoptions } from './worldoptions.js'
 
-const keybindsBtn = Btn(texts.keybinds())
-keybindsBtn.disabled = true
 const spage = Btn(texts.pause.server_page(), () => {
 	if(ws.opts) return worldoptions(ws)
 	const { protocol, host } = new URL(ws.url)
 	open(protocol.replace('ws','http')+'//'+host,'_blank')
-})
+}, 'small')
 const pauseui = UI('menu',
 	Label(texts.pause.name()),
 	Btn(texts.pause.resume(), hideUI),
-	Row(Btn(texts.options.general(), optionsScreen), spage),
-	Row(Btn(texts.options.controls(), controlsScreen), keybindsBtn),
+	Row(Btn(texts.options.general(), optionsScreen, 'small'), spage),
+	Row(Btn(texts.options.controls(), controlsScreen, 'small'), Btn(texts.keybinds(), null, 'small disabled')),
 	Btn(texts.connection.disconnect(), serverlist),
 	Div('',
 		Label(texts.pause.credits()).css({height:'unset',cursor:'pointer',lineHeight:'unset'}).attr('onclick', "window.open('https://github.com/open-mc/client/blob/main/CREDITS.md','_blank')"),